Salience/scripts/requirements-cam.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.6.15=py38h06a4308_0
cffi=1.15.1=py38h74dc2b5_0
colorama=0.4.5=pyhd8ed1ab_0
cycler=0.11.0=pyhd8ed1ab_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
future=0.18.2=py38_1
giflib=5.2.1=h7b6447c_0
glob2=0.7=py_0
icu=67.1=he1b5a44_0
intel-openmp=2021.4.0=h06a4308_3561
ipython=8.4.0=py38h06a4308_0
jedi=0.18.1=py38h06a4308_1
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py38h295c915_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.2.2=1
matplotlib-base=3.2.2=py38h5d868c9_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.3=h5eee18b_3
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numexpr=2.8.3=py38h807cd23_0
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py38h6a678d5_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.9=pyhd8ed1ab_0
python=3.8.13=haa1d7c7_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.10.2=cpu_py38hfa7516b_0
pytz=2022.1=py38h06a4308_0
readline=8.2=h5eee18b_0
scikit-learn=1.1.1=py38h6a678d5_0
scipy=1.7.3=py38hc147768_0
setuptools=63.4.1=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
stack_data=0.2.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchcam=0.3.2=pyhd8ed1ab_0
torchvision=0.11.3=cpu_py38h164cc8f_2
tornado=6.1=py38h0a891b7_3
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py38h06a4308_0
typing_extensions=4.3.0=py38h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=h5eee18b_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
